requirement.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tomli==1.2.1
uritemplate==3.0.1
urllib3==1.26.6
